+++ packages/akashic-cli-serve/src/common/types/SandboxConfig.ts
@@ -6,8 +6,15 @@ export interface SandboxConfig {
 	autoSendEvents?: string;
 	autoSendEventName?: string;
 	backgroundImage?: string;
+	backgroundColor?: string;
 	showMenu?: boolean;
 	events?: { [name: string]: any };
 	arguments?: { [name: string]: any };
 	externalAssets?: (string | RegExp)[] | null;
+	client?: {
+		external?: {
+			[key: string]: string;  // 値は各プラグインの scriptPath
+		};
+	};
+
 }